+++ v1_10/requirements.txt
@@ -1,0 +1,3 @@
+pydantic==1.10.13
+srsly==2.4.8
+pytest-benchmark==4.0.0
